euler5

+++ src/main/java/com/company/Main.java
@@ -13,7 +13,7 @@ class Main {
         org.apache.commons.lang.time.StopWatch stopwatch  = new StopWatch();
         stopwatch.start();
 
-        int[] numbers = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20};
+        long[] numbers = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20};
 
         System.out.println(computeLCM(numbers));
         stopwatch.stop(); // optional
@@ -22,16 +22,16 @@ class Main {
 
     }
 
-    public static int computeLCM(int[] numbers) {
-        int lcm = numbers[0];
+    public static long computeLCM(long[] numbers) {
+        long lcm = numbers[0];
         for (int i = 1; i < numbers.length; i++) {
             lcm = (lcm * numbers[i]) / computeGCD(lcm, numbers[i]);
-        }
+         }
         return lcm;
     }
 
 
-    public static int computeGCD(int a, int b) {
+    public static long computeGCD(long a, long b) {
         if (b == 0) {
             return a;
         }
